+++ frontend/src/components/editor/cell/code/cell-editor.tsx
@@ -37,7 +37,7 @@ import { invariant } from "@/utils/invariant";
 import { mergeRefs } from "@/utils/mergeRefs";
 import { AiCompletionEditor } from "../../ai/ai-completion-editor";
 import { HideCodeButton } from "../../code/readonly-python-code";
-import { useCellEditorNavigationProps } from "../../focus/focus";
+import { useCellEditorNavigationProps } from "../../navigation/navigation";
 import { useDeleteCellCallback } from "../useDeleteCell";
 import { useSplitCellCallback } from "../useSplitCell";
 import { LanguageToggles } from "./language-toggle";
