+++ frontend/src/components/editor/navigation/__tests__/clipboard.test.ts
