+++ frontend/src/components/editor/navigation/clipboard.ts
