+++ frontend/src/components/editor/navigation/focus-utils.ts
