+++ frontend/src/components/editor/navigation/state.ts
